tests employee

+++ src/main/java/com/allocation/manager/service/impl/EmployeeServiceImpl.java
@@ -3,6 +3,7 @@ package com.allocation.manager.service.impl;
 import com.allocation.manager.model.Employee;
 import com.allocation.manager.repository.EmployeeRepository;
 import com.allocation.manager.service.IEmployeeService;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -37,6 +38,9 @@ public class EmployeeServiceImpl implements IEmployeeService {
 
     @Override
     public void deleteEmployee(UUID employeeId) {
+        if(!employeeRepository.existsById(employeeId)){
+            throw new EntityNotFoundException("Nenhum colaborador foi encontrado com o ID fornecido: " + employeeId);
+        }
         employeeRepository.deleteById(employeeId);
     }
